requirements.txt
# Core MCP dependencies
mcp>=1.0.0
langchain>=1.0.0
langchain-mcp-adapters>=0.1.9
langchain-google-genai>=1.0.0
langgraph>=0.0.30

# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.1

# Utilities
python-dotenv>=1.1.1
aiofiles>=23.2.1
pydantic-settings>=2.2.1
requests>=2.32.3

# Performance (explicitly include these since uvicorn[standard] might not work in all environments)
httptools>=0.5.0
uvloop>=0.17.0; sys_platform != "win32"

#redis 
redis==5.0.1
sentence-transformers==2.2.2
numpy>=1.26.0

# Instance management dependencies
PyJWT>=2.8.0
pycryptodome>=3.19.0

google-cloud-secret-manager>=2.16.0